Support for external git dependencies with selectable commit id / branch / tag name

+++ src/main/java/com/github/zhgzhg/drizzle/utils/file/FileUtils.java
@@ -2,6 +2,14 @@ package com.github.zhgzhg.drizzle.utils.file;
 
 import com.github.zhgzhg.drizzle.utils.log.LogProxy;
 import com.github.zhgzhg.drizzle.utils.text.TextUtils;
+import org.eclipse.jgit.api.CloneCommand;
+import org.eclipse.jgit.api.CreateBranchCommand;
+import org.eclipse.jgit.api.Git;
+import org.eclipse.jgit.api.errors.GitAPIException;
+import org.eclipse.jgit.api.errors.RefNotFoundException;
+import org.eclipse.jgit.lib.Ref;
+import org.eclipse.jgit.lib.Repository;
+import org.eclipse.jgit.storage.file.FileRepositoryBuilder;
 
 import java.io.File;
 import java.io.FileNotFoundException;
@@ -23,8 +31,30 @@ import java.util.stream.Stream;
 public class FileUtils<T> {
     private final LogProxy<T> logProxy;
 
+    public static class RepoLibDir {
+        public final String url;
+        public final String revision;
+        public final String resultingLibNameDir;
+        public final File dir;
+
+        public RepoLibDir(String url, String revision, String libName, File dir) {
+            this.url = url;
+            this.revision = revision;
+            this.resultingLibNameDir = libName;
+            this.dir = dir;
+        }
+    }
+
     public FileUtils(LogProxy<T> logProxy) {
         this.logProxy = logProxy;
+    }
+
+    public RepoLibDir downloadGit(String url, String desiredLibNameDir, String tempFileNamePostfix) {
+        URL realUrl = TextUtils.toURL(url, logProxy::cliErrorln);
+        if (realUrl != null) {
+            return downloadGit(realUrl, desiredLibNameDir, tempFileNamePostfix);
+        }
+        return null;
     }
 
     public File downloadZip(String url, String tempFileNamePostfix) {
@@ -33,6 +63,63 @@ public class FileUtils<T> {
             return downloadZip(realUrl, tempFileNamePostfix);
         }
         return null;
+    }
+
+    public RepoLibDir downloadGit(URL url, String desiredLibNameDir, String tempDirNamePostfix) {
+
+        File tempDir = null;
+        try {
+            tempDir = Files.createTempDirectory("ard-drizzle-ext" + (tempDirNamePostfix != null ? tempDirNamePostfix : "")).toFile();
+            tempDir.deleteOnExit();
+
+            String revision = (TextUtils.isNotNullOrBlank(url.getRef()) ? url.getRef() : null);
+            String libName = desiredLibNameDir;
+            String uri = TextUtils.rtrim(url.toExternalForm(), (revision != null ? "#" + revision : null));
+
+            CloneCommand cloneCommand = Git.cloneRepository().setURI(uri).setCloneSubmodules(true).setCloneAllBranches(true);
+
+            File subLibDir = new File(tempDir, desiredLibNameDir);
+            if (!subLibDir.mkdir()) {
+                throw new IOException("Cannot create temp. directory " + libName);
+            }
+
+            cloneCommand.setDirectory(subLibDir);
+
+            if (revision != null) {
+                try (Git result = cloneCommand.call();
+                     Repository repository = result.getRepository()) {
+
+                    Ref ref = repository.findRef(revision);
+                    if (ref != null) {
+                        result.checkout().setName(ref.getName()).call();
+                    } else {
+                        try {
+                            result.checkout()
+                                    .setUpstreamMode(CreateBranchCommand.SetupUpstreamMode.SET_UPSTREAM)
+                                    .setName(revision)
+                                    .call();
+                        } catch (RefNotFoundException fnfe) {
+                            result.checkout()
+                                    .setUpstreamMode(CreateBranchCommand.SetupUpstreamMode.SET_UPSTREAM)
+                                    .setName("origin/" + revision)
+                                    .call();
+                        }
+                    }
+                }
+            }
+
+            return new RepoLibDir(uri, revision, libName, subLibDir);
+        } catch (IOException | GitAPIException e) {
+            this.logProxy.cliError("Failed downloading the external git repo library %s - %s%n", url, e.getMessage());
+            if (tempDir != null) {
+                try {
+                    removeDir(tempDir);
+                } catch (Exception ex) {
+                    // don't care
+                }
+            }
+            return null;
+        }
     }
 
     public File downloadZip(URL url, String tempFileNamePostfix) {
@@ -73,14 +160,46 @@ public class FileUtils<T> {
                     e.printStackTrace(this.logProxy.stderr());
                     Thread.currentThread().interrupt();
                 } finally {
-                    try {
-                        Files.deleteIfExists(tempFile.toPath());
-                    } catch (IOException e) {
-                        this.logProxy.cliErrorln(e);
-                    }
+                    removeFile(tempFile);
                 }
             }).start();
         }
+    }
+
+    public void delayedDirRemoval(long millis, File tempDir) {
+        if (tempDir != null) {
+            new Thread(() -> {
+                try {
+                    Thread.sleep(millis);
+                } catch (InterruptedException e) {
+                    e.printStackTrace(this.logProxy.stderr());
+                    Thread.currentThread().interrupt();
+                } finally {
+                    removeDir(tempDir);
+                }
+            }).start();
+        }
+    }
+
+    private void removeFile(File tempFile) {
+        if (tempFile != null) {
+            try {
+                Files.deleteIfExists(tempFile.toPath());
+            } catch (IOException e) {
+                this.logProxy.cliErrorln(e);
+            }
+        }
+    }
+
+    private void removeDir(File dir) {
+        File[] contents = dir.listFiles();
+        if (contents != null) {
+            for (File f : contents) {
+                removeDir(f);
+            }
+        }
+        dir.delete();
+        //removeFile(dir);
     }
 
     public List<Path> listJarsInDir(Path inPath) {
